Update metadata for F-Droid mainline repo
Updated: Metadata for F-Droid
Added: Screenshots for F-Droid

There were previous tries to get screenshots to
show up in F-Droid which didn't succeed.
This new metadata is tested in a private repo
and works in newer clients.
THe problem is that the F-Droid Client doesn't
recognize localized folders as expected.
It is said to look at e.g. "de" folder if
"de-DE" doesn't exist. This is suggested
by F-Droid Client Code, but it doesn't work.

Signed-off-by: Gregor Santner <gsantner@mailbox.org>